Restructure tests to avoid installing knative more than once
It's silly to re-install both serving *and* eventing only to run a few
more serving tests (but no eventing ones!!!). We should run all the
e2e tests in a single installation of each component.

TODO: Refactor the subtests to allow running single tests by adding a
"-run" arg to TEST_ARGS. You can do this, currently:
  $ TEST_ARGS="-run Serving" SCALE_UP=0 make test-e2e
But we'd like to be able to do this, too:
  $ TEST_ARGS="-run Serving/downstream/https" make test-e2e
That requires a guard to ensure the CR is installed.
